Move autobumper out of experimental directory
