--- experiment/autobumper/cloudbuild.yaml
@@ -1,26 +1,0 @@
-steps:
-  - name: golang:$_GO_VERSION
-    args:
-    - go
-    - build
-    - -mod=readonly
-    - -a
-    - -installsuffix=cgo
-    - -o=./experiment/autobumper/generic_autobump
-    - ./experiment/autobumper/main.go
-    env:
-    - CGO_ENABLED=0
-    - GOOS=linux
-    - GOARCH=amd64
-  - name: gcr.io/cloud-builders/docker
-    args:
-    - build
-    - --tag=gcr.io/$PROJECT_ID/generic_autobump:$_GIT_TAG
-    - --tag=gcr.io/$PROJECT_ID/generic_autobump:latest
-    - ./experiment/autobumper
-substitutions:
-  _GIT_TAG: '12345'
-  _GO_VERSION: 1.13-alpine
-images:
-  - 'gcr.io/$PROJECT_ID/generic_autobump:$_GIT_TAG'
-  - 'gcr.io/$PROJECT_ID/generic_autobump:latest'
