+++ hack/autobump.sh
@@ -17,4 +17,4 @@ set -o nounset
 set -o errexit
 set -o pipefail
 
-bazel run //experiment/autobumper -- %@
+bazel run //prow/cmd/generic-autobumper -- %@
