+++ prow/cmd/generic-autobumper/BUILD.bazel
@@ -1,22 +1,22 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 load("//prow:def.bzl", "prow_image")
 
-NAME = "autobumper"
+NAME = "generic-autobumper"
 
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "k8s.io/test-infra/experiment/autobumper",
+    importpath = "k8s.io/test-infra/prow/cmd/generic-autobumper",
     visibility = ["//visibility:private"],
     deps = [
-        "//experiment/autobumper/bumper:go_default_library",
+        "//prow/cmd/generic-autobumper/bumper:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@in_gopkg_yaml_v2//:go_default_library",
     ],
 )
 
 go_binary(
-    name = "autobumper",
+    name = "generic-autobumper",
     embed = [":go_default_library"],
     visibility = ["//visibility:public"],
 )
@@ -32,7 +32,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//experiment/autobumper/bumper:all-srcs",
+        "//prow/cmd/generic-autobumper/bumper:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
@@ -43,6 +43,6 @@ prow_image(
     base = "@git-base//image",
     component = NAME,
     directory = "/",
-    files = [":autobumper"],
+    files = [":generic-autobumper"],
     visibility = ["//visibility:public"],
 )
