+++ prow/cmd/generic-autobumper/bumper/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["bumper.go"],
-    importpath = "k8s.io/test-infra/experiment/autobumper/bumper",
+    importpath = "k8s.io/test-infra/prow/cmd/generic-autobumper/bumper",
     visibility = ["//visibility:public"],
     deps = [
         "//experiment/image-bumper/bumper:go_default_library",
