+++ prow/cmd/generic-autobumper/bumper/bumper_test.go
