+++ prow/cmd/generic-autobumper/main.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"k8s.io/test-infra/experiment/autobumper/bumper"
+	"k8s.io/test-infra/prow/cmd/generic-autobumper/bumper"
 
 	"gopkg.in/yaml.v2"
 )
